changed file organisation, split up cwl yaml parts into separate files
