+++ src/cwl/clt.rs
@@ -1,12 +1,14 @@
 use super::{
     execution::runner::run_command,
-    types::{CWLType, Directory, EnvironmentDef, File},
+    inputs::{deserialize_inputs, CommandInputParameter, CommandLineBinding},
+    outputs::{deserialize_outputs, CommandOutputParameter},
+    requirements::{deserialize_requirements, DockerRequirement, Requirement},
+    types::{DefaultValue, Entry},
 };
 use crate::io::resolve_path;
 use core::fmt;
-use serde::{Deserialize, Deserializer, Serialize};
-use serde_yml::{Mapping, Value};
-use std::{collections::HashMap, error::Error, fmt::Display, vec};
+use serde::{Deserialize, Serialize};
+use std::{error::Error, fmt::Display};
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -178,399 +180,4 @@ impl Default for Command {
     fn default() -> Self {
         Command::Single(String::new())
     }
-}
-
-#[derive(Serialize, Deserialize, Debug, Default, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct CommandInputParameter {
-    #[serde(default)]
-    pub id: String,
-    pub type_: CWLType,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub default: Option<DefaultValue>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub input_binding: Option<CommandLineBinding>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<String>,
-}
-
-impl CommandInputParameter {
-    pub fn with_id(mut self, id: &str) -> Self {
-        self.id = id.to_string();
-        self
-    }
-
-    pub fn with_type(mut self, t: CWLType) -> Self {
-        self.type_ = t;
-        self
-    }
-
-    pub fn with_default_value(mut self, f: DefaultValue) -> Self {
-        self.default = Some(f);
-        self
-    }
-
-    pub fn with_binding(mut self, binding: CommandLineBinding) -> Self {
-        self.input_binding = Some(binding);
-        self
-    }
-}
-
-#[derive(Serialize, Debug, PartialEq, Clone)]
-#[serde(untagged)]
-pub enum DefaultValue {
-    File(File),
-    Directory(Directory),
-    Any(serde_yml::Value),
-}
-
-impl DefaultValue {
-    pub fn as_value_string(&self) -> String {
-        match self {
-            DefaultValue::File(item) => item.location.clone(),
-            DefaultValue::Directory(item) => item.location.clone(),
-            DefaultValue::Any(value) => match value {
-                serde_yml::Value::Bool(_) => String::from(""), // do not remove!
-                _ => serde_yml::to_string(value).unwrap().trim_end().to_string(),
-            },
-        }
-    }
-
-    pub fn has_matching_type(&self, cwl_type: &CWLType) -> bool {
-        matches!(
-            (self, cwl_type),
-            (DefaultValue::File(_), CWLType::File) | (DefaultValue::Directory(_), CWLType::Directory) | (DefaultValue::Any(_), _)
-        )
-    }
-}
-
-impl<'de> Deserialize<'de> for DefaultValue {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let value: Value = Deserialize::deserialize(deserializer)?;
-
-        let location = value.get("location").or_else(|| value.get("path")).and_then(Value::as_str);
-
-        if let Some(location_str) = location {
-            let secondary_files = value
-                .get("secondaryFiles")
-                .map(|v| serde_yml::from_value(v.clone()))
-                .transpose()
-                .map_err(serde::de::Error::custom)?;
-
-            let basename = value
-                .get("basename")
-                .map(|v| serde_yml::from_value(v.clone()))
-                .transpose()
-                .map_err(serde::de::Error::custom)?;
-
-            match value.get("class").and_then(Value::as_str) {
-                Some("File") => {
-                    let format = value
-                        .get("format")
-                        .map(|v| serde_yml::from_value(v.clone()))
-                        .transpose()
-                        .map_err(serde::de::Error::custom)?;
-                    let mut item = File::from_location(&location_str.to_string());
-                    item.secondary_files = secondary_files;
-                    item.basename = basename;
-                    item.format = format;
-                    Ok(DefaultValue::File(item))
-                }
-                Some("Directory") => {
-                    let mut item = Directory::from_location(&location_str.to_string());
-                    item.secondary_files = secondary_files;
-                    item.basename = basename;
-                    Ok(DefaultValue::Directory(item))
-                }
-                _ => Ok(DefaultValue::Any(value)),
-            }
-        } else {
-            Ok(DefaultValue::Any(value))
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Default, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct CommandLineBinding {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub prefix: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub position: Option<isize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value_from: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub shell_quote: Option<bool>,
-}
-
-impl CommandLineBinding {
-    pub fn with_prefix(mut self, prefix: &String) -> Self {
-        self.prefix = Some(prefix.to_string());
-        self
-    }
-
-    pub fn with_position(mut self, position: isize) -> Self {
-        self.position = Some(position);
-        self
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Default, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct CommandOutputParameter {
-    #[serde(default)]
-    pub id: String,
-    pub type_: CWLType,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub output_binding: Option<CommandOutputBinding>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<String>,
-}
-
-impl CommandOutputParameter {
-    pub fn with_id(mut self, id: &str) -> Self {
-        self.id = id.to_string();
-        self
-    }
-    pub fn with_type(mut self, type_: CWLType) -> Self {
-        self.type_ = type_;
-        self
-    }
-    pub fn with_binding(mut self, binding: CommandOutputBinding) -> Self {
-        self.output_binding = Some(binding);
-        self
-    }
-}
-
-fn deserialize_inputs<'de, D>(deserializer: D) -> Result<Vec<CommandInputParameter>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value: Value = Deserialize::deserialize(deserializer)?;
-
-    let parameters = match value {
-        Value::Sequence(seq) => seq
-            .into_iter()
-            .map(|item| {
-                let param: CommandInputParameter = serde_yml::from_value(item).map_err(serde::de::Error::custom)?;
-                Ok(param)
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        Value::Mapping(map) => map
-            .into_iter()
-            .map(|(key, value)| {
-                let id = key.as_str().ok_or_else(|| serde::de::Error::custom("Expected string key"))?;
-                let param = match value {
-                    Value::String(type_str) => {
-                        let type_ = serde_yml::from_value::<CWLType>(Value::String(type_str)).map_err(serde::de::Error::custom)?;
-                        CommandInputParameter::default().with_id(id).with_type(type_)
-                    }
-                    _ => {
-                        let mut param: CommandInputParameter = serde_yml::from_value(value).map_err(serde::de::Error::custom)?;
-                        param.id = id.to_string();
-                        param
-                    }
-                };
-
-                Ok(param)
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        _ => return Err(serde::de::Error::custom("Expected sequence or mapping for inputs")),
-    };
-
-    Ok(parameters)
-}
-
-fn deserialize_outputs<'de, D>(deserializer: D) -> Result<Vec<CommandOutputParameter>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value: Value = Deserialize::deserialize(deserializer)?;
-
-    let parameters = match value {
-        Value::Sequence(seq) => seq
-            .into_iter()
-            .map(|item| {
-                let param: CommandOutputParameter = serde_yml::from_value(item).map_err(serde::de::Error::custom)?;
-                Ok(param)
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        Value::Mapping(map) => map
-            .into_iter()
-            .map(|(key, value)| {
-                let id = key.as_str().ok_or_else(|| serde::de::Error::custom("Expected string key"))?;
-                let mut param: CommandOutputParameter = serde_yml::from_value(value).map_err(serde::de::Error::custom)?;
-                param.id = id.to_string();
-                Ok(param)
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        _ => return Err(serde::de::Error::custom("Expected sequence or mapping for outputs")),
-    };
-
-    Ok(parameters)
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct CommandOutputBinding {
-    pub glob: String,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(tag = "class")]
-pub enum Requirement {
-    InitialWorkDirRequirement(InitialWorkDirRequirement),
-    DockerRequirement(DockerRequirement),
-    ResourceRequirement(ResourceRequirement),
-    EnvVarRequirement(EnvVarRequirement),
-    ShellCommandRequirement,
-}
-
-fn deserialize_requirements<'de, D>(deserializer: D) -> Result<Option<Vec<Requirement>>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value: Option<Value> = Deserialize::deserialize(deserializer)?;
-    if value.is_none() {
-        return Ok(None);
-    }
-
-    let value = value.unwrap();
-
-    let parameters = match value {
-        Value::Sequence(seq) => seq
-            .into_iter()
-            .map(|item| {
-                let param: Requirement = serde_yml::from_value(item).map_err(serde::de::Error::custom)?;
-                Ok(param)
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        Value::Mapping(map) => map
-            .into_iter()
-            .map(|(key, value)| {
-                let class = key.as_str().ok_or_else(|| serde::de::Error::custom("Expected string key"))?;
-                let mut modified_value = value;
-                let new_map = match modified_value {
-                    Value::Mapping(ref mut inner_map) => {
-                        inner_map.insert(Value::String("class".to_string()), Value::String(class.to_string()));
-                        inner_map.clone()
-                    }
-                    _ => {
-                        let mut map = Mapping::new();
-                        map.insert(Value::String("class".to_string()), Value::String(class.to_string()));
-                        map
-                    }
-                };
-                let param: Requirement = serde_yml::from_value(Value::Mapping(new_map)).map_err(serde::de::Error::custom)?;
-                Ok(param)
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        _ => return Err(serde::de::Error::custom("Expected sequence or mapping for outputs")),
-    };
-
-    Ok(Some(parameters))
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct InitialWorkDirRequirement {
-    pub listing: Vec<Listing>,
-}
-
-impl InitialWorkDirRequirement {
-    pub fn from_file(filename: &str) -> Self {
-        InitialWorkDirRequirement {
-            listing: vec![Listing {
-                entryname: filename.to_string(),
-                entry: Entry::from_file(filename),
-            }],
-        }
-    }
-    pub fn from_contents(entryname: &str, contents: &str) -> Self {
-        InitialWorkDirRequirement {
-            listing: vec![Listing {
-                entryname: entryname.to_string(),
-                entry: Entry::Source(contents.to_string()),
-            }],
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub enum DockerRequirement {
-    DockerPull(String),
-    #[serde(untagged)]
-    DockerFile {
-        #[serde(rename = "dockerFile")]
-        docker_file: Entry,
-        #[serde(rename = "dockerImageId")]
-        docker_image_id: String,
-    },
-}
-
-impl DockerRequirement {
-    pub fn from_file(filename: &str, tag: &str) -> Self {
-        DockerRequirement::DockerFile {
-            docker_file: Entry::from_file(filename),
-            docker_image_id: tag.to_string(),
-        }
-    }
-    pub fn from_pull(image_id: &str) -> Self {
-        DockerRequirement::DockerPull(image_id.to_string())
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct ResourceRequirement {
-    pub cores_min: Option<i32>,
-    pub cores_max: Option<i32>,
-    pub ram_min: Option<i32>,
-    pub ram_max: Option<i32>,
-    pub tmpdir_min: Option<i32>,
-    pub tmpdir_max: Option<i32>,
-    pub outdir_min: Option<i32>,
-    pub outdir_max: Option<i32>,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct EnvVarRequirement {
-    pub env_def: EnviromentDefs,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
-#[serde(untagged)]
-pub enum EnviromentDefs {
-    Vec(Vec<EnvironmentDef>),
-    Map(HashMap<String, String>),
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct Listing {
-    pub entryname: String,
-    pub entry: Entry,
-}
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(untagged)]
-pub enum Entry {
-    Source(String),
-    Include(Include),
-}
-
-impl Entry {
-    pub fn from_file(path: &str) -> Entry {
-        Entry::Include(Include { include: path.to_string() })
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
-pub struct Include {
-    #[serde(rename = "$include")]
-    pub include: String,
 }
