+++ src/cwl/mod.rs
@@ -1,10 +1,13 @@
 pub mod clt;
+pub mod execution;
 pub mod format;
+pub mod inputs;
 pub mod loader;
+pub mod outputs;
+pub mod parser;
+pub mod requirements;
 pub mod types;
 pub mod wf;
-pub mod execution;
-pub mod parser;
 
 #[cfg(test)]
 mod cwl_tests;
