+++ src/cwl/types.rs
@@ -1,5 +1,7 @@
-use super::clt::DefaultValue;
-use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+
+use serde::{Deserialize, Deserializer, Serialize};
+use serde_yml::Value;
 
 #[derive(Serialize, Deserialize, Debug, Default, PartialEq, Clone)]
 #[serde(rename_all = "camelCase")]
@@ -16,6 +18,83 @@ pub enum CWLType {
     File,
     #[serde(rename = "Directory")]
     Directory,
+}
+
+#[derive(Serialize, Debug, PartialEq, Clone)]
+#[serde(untagged)]
+pub enum DefaultValue {
+    File(File),
+    Directory(Directory),
+    Any(serde_yml::Value),
+}
+
+impl DefaultValue {
+    pub fn as_value_string(&self) -> String {
+        match self {
+            DefaultValue::File(item) => item.location.clone(),
+            DefaultValue::Directory(item) => item.location.clone(),
+            DefaultValue::Any(value) => match value {
+                serde_yml::Value::Bool(_) => String::from(""), // do not remove!
+                _ => serde_yml::to_string(value).unwrap().trim_end().to_string(),
+            },
+        }
+    }
+
+    pub fn has_matching_type(&self, cwl_type: &CWLType) -> bool {
+        matches!(
+            (self, cwl_type),
+            (DefaultValue::File(_), CWLType::File) | (DefaultValue::Directory(_), CWLType::Directory) | (DefaultValue::Any(_), _)
+        )
+    }
+}
+
+impl<'de> Deserialize<'de> for DefaultValue {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let value: Value = Deserialize::deserialize(deserializer)?;
+
+        let location = value.get("location").or_else(|| value.get("path")).and_then(Value::as_str);
+
+        if let Some(location_str) = location {
+            let secondary_files = value
+                .get("secondaryFiles")
+                .map(|v| serde_yml::from_value(v.clone()))
+                .transpose()
+                .map_err(serde::de::Error::custom)?;
+
+            let basename = value
+                .get("basename")
+                .map(|v| serde_yml::from_value(v.clone()))
+                .transpose()
+                .map_err(serde::de::Error::custom)?;
+
+            match value.get("class").and_then(Value::as_str) {
+                Some("File") => {
+                    let format = value
+                        .get("format")
+                        .map(|v| serde_yml::from_value(v.clone()))
+                        .transpose()
+                        .map_err(serde::de::Error::custom)?;
+                    let mut item = File::from_location(&location_str.to_string());
+                    item.secondary_files = secondary_files;
+                    item.basename = basename;
+                    item.format = format;
+                    Ok(DefaultValue::File(item))
+                }
+                Some("Directory") => {
+                    let mut item = Directory::from_location(&location_str.to_string());
+                    item.secondary_files = secondary_files;
+                    item.basename = basename;
+                    Ok(DefaultValue::Directory(item))
+                }
+                _ => Ok(DefaultValue::Any(value)),
+            }
+        } else {
+            Ok(DefaultValue::Any(value))
+        }
+    }
 }
 
 pub trait PathItem {
@@ -99,6 +178,38 @@ impl PathItem for Directory {
     fn location(&self) -> &String {
         &self.location
     }
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[serde(untagged)]
+pub enum EnviromentDefs {
+    Vec(Vec<EnvironmentDef>),
+    Map(HashMap<String, String>),
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct Listing {
+    pub entryname: String,
+    pub entry: Entry,
+}
+#[derive(Serialize, Deserialize, Debug, PartialEq)]
+#[serde(untagged)]
+pub enum Entry {
+    Source(String),
+    Include(Include),
+}
+
+impl Entry {
+    pub fn from_file(path: &str) -> Entry {
+        Entry::Include(Include { include: path.to_string() })
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+pub struct Include {
+    #[serde(rename = "$include")]
+    pub include: String,
 }
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
